+++ src/fs/file.rs
@@ -1,9 +1,6 @@
 use crate::{
     collections::vec::NSTDVec,
-    core::{
-        def::{NSTDAny, NSTDChar},
-        slice::NSTDSlice,
-    },
+    core::{def::NSTDAny, slice::NSTDSlice, str::NSTDStr},
     io::{
         input_stream::NSTDInputStream, io_stream::NSTDIOStream, output_stream::NSTDOutputStream,
         stream::NSTDStream,
@@ -11,7 +8,6 @@ use crate::{
     string::NSTDString,
 };
 use std::{
-    ffi::CStr,
     fs::File,
     io::{prelude::*, BufReader},
 };
@@ -45,7 +41,7 @@ pub struct NSTDFile {
 
 /// Opens a file and returns the file stream. Files must be closed.
 /// Parameters:
-///     `const NSTDChar *const name` - The name of the file.
+///     `const NSTDStr *const name` - The name of the file.
 ///     `const NSTDUSize mask` - Bit mask defining how to open the file.
 ///         - Bit 1 - Create the file if it doesn't exist. Write bit must be set for this to work.
 ///         - Bit 2 - Read from the file.
@@ -54,8 +50,8 @@ pub struct NSTDFile {
 ///         - Bit 5 - Truncate the file.
 /// Returns: `NSTDFile file` - The file stream.
 #[cfg_attr(feature = "clib", no_mangle)]
-pub unsafe extern "C" fn nstd_fs_file_open(name: *const NSTDChar, mask: usize) -> NSTDFile {
-    let stream = NSTDStream { errc: 0 };
+pub unsafe extern "C" fn nstd_fs_file_open(name: &NSTDStr, mask: usize) -> NSTDFile {
+    let stream = NSTDStream::default();
     let io_stream = NSTDIOStream {
         input_stream: NSTDInputStream {
             stream,
@@ -70,7 +66,7 @@ pub unsafe extern "C" fn nstd_fs_file_open(name: *const NSTDChar, mask: usize) -
             write: Some(fs_ostream_write),
         },
     };
-    if let Ok(name) = CStr::from_ptr(name).to_str() {
+    if let Ok(name) = std::str::from_utf8(name.bytes.as_byte_slice()) {
         if let Ok(f) = File::options()
             .create(mask & NSTD_FS_FILE_CREATE != 0)
             .read(mask & NSTD_FS_FILE_READ != 0)